Move HogQL insights to their own folder
